Make PhoneNumbers.update_instance consistent with PhoneNumber.update
A recent change to PhoneNumber.update allowed passing "status_callback"
as "status_callback_url" for convenience. This change adds this
conveniece for PhoneNumbers.update_instance for consistency. Also,
kwargs is now copied to avoid side effects and status_callback_url is now
removed from kwargs as it is only relevant to update and
update_instance.